server/requirements.txt
certifi==2024.12.14
distlib==0.3.9
filelock==3.16.1
packaging==24.2
pipenv==2024.4.0
platformdirs==4.3.6
python-dotenv==1.0.1
virtualenv==20.28.0
